+++ src/Umbraco.Web.UI.Client/src/packages/media/imaging/components/index.ts
@@ -1,3 +1,3 @@
-export * from './imaging-thumbnail.element.js';
 export * from './thumbnail.element.js';
+export * from './imaging-thumbnail.element.js';
 export * from './media-image.element.js';
