+++ src/Umbraco.Web.UI.Client/src/packages/media/imaging/stories/working-with-media-images.mdx
@@ -16,8 +16,8 @@ handle authentication, HMAC-signed URLs, lazy loading and caching for you.
 | `<umb-thumbnail>` | ✅ server-side resize & crop | ✅ configurable | You want a sized thumbnail — cards, pickers, collections, block grid items. **Recommended default.** |
 | `<umb-media-image>` | ❌ original file | ✅ | You need the original, full-size image without resizing. |
 
-> `<umb-imaging-thumbnail>` is the original name of `<umb-thumbnail>` and remains registered for backwards
-> compatibility. Prefer `<umb-thumbnail>` in new code.
+> `<umb-imaging-thumbnail>` is a deprecated alias of `<umb-thumbnail>`, kept for backwards compatibility
+> and scheduled for removal in Umbraco 19. Use `<umb-thumbnail>` in new code.
 
 ## Displaying a thumbnail
