+++ src/Umbraco.Web.UI.Client/src/packages/media/media/collection/item/media-collection-item-card.element.ts
@@ -72,10 +72,10 @@ export class UmbMediaCollectionItemCardElement extends UmbLitElement implements 
 				?disabled=${this.disabled}
 				@selected=${this.#onSelected}
 				@deselected=${this.#onDeselected}>
-				<umb-imaging-thumbnail
+				<umb-thumbnail
 					.unique=${this.item.unique}
 					alt=${ifDefined(this.item.name)}
-					icon=${ifDefined(this.item.icon)}></umb-imaging-thumbnail>
+					icon=${ifDefined(this.item.icon)}></umb-thumbnail>
 				<slot name="actions" slot="actions"></slot>
 			</uui-card-media>
 		`;
