+++ src/Umbraco.Web.UI.Client/src/packages/media/media/components/input-rich-media/input-rich-media.element.ts
@@ -413,11 +413,11 @@ export class UmbInputRichMediaElement extends UmbFormControlMixin<
 
 		return html`
 			<uui-card-media id=${item.unique} title=${item.name} name=${item.name} .href=${href} ?readonly=${this.readonly}>
-				<umb-imaging-thumbnail
+				<umb-thumbnail
 					.unique=${item.media}
 					.alt=${item.name}
 					.icon=${item.icon ?? 'icon-picture'}
-					.externalLoading=${item.isLoading ?? false}></umb-imaging-thumbnail>
+					.externalLoading=${item.isLoading ?? false}></umb-thumbnail>
 
 				${this.#renderIsTrashed(item)} ${this.#renderActions(item)}
 			</uui-card-media>
